refactor: extract Gallery component from HomePage
