+++ src/components/UI/Container/Container.module.scss
@@ -3,13 +3,9 @@
 .container {
   margin: 0 auto;
   max-width: 1280px;
-  padding: 40px 50px;
+  padding: 20px 50px;
 
   @include for(Tablet) {
-    padding: 20px;
-  }
-
-  @include for(TabletP) {
     padding: 15px;
   }
